cyfuture/requirements.txt
python-dotenv
google-generativeai
PyMuPDF
streamlit
